refactor: move LSP service owners out of core
