+++ src/crates/adapters/ai-adapters/src/cli_credentials/gemini.rs
