+++ src/crates/services/services-core/AGENTS.md
@@ -2,14 +2,15 @@
 
 Scope: this guide applies to `src/crates/services/services-core`.
 
-`bitfun-services-core` owns platform-neutral service DTOs and helpers that can
+`bitfun-services-core` owns cross-platform service DTOs and helpers that can
 compile without the full product runtime. It also owns generic local filesystem
-operations/tree/search/listing primitives, LSP plugin registry and command-target
-mapping rules, session storage layout helpers, turn file indexing/deletion,
-metadata store CRUD/index rebuild, metadata construction/counter/index/field
-mutation rules, lineage/branch metadata shaping, and reusable JSON file IO;
-product crates may layer remote workspace routing or legacy error mapping outside
-this crate.
+operations/tree/search/listing primitives, reusable LSP registry/package
+loading/protocol/project detection/config watching/debounce/process-manager
+helpers, session storage layout helpers, turn file indexing/deletion, metadata
+store CRUD/index rebuild, metadata construction/counter/index/field mutation
+rules, lineage/branch metadata shaping, and reusable JSON file IO; product
+crates may layer remote workspace routing or legacy error mapping outside this
+crate.
 
 ## Guardrails
 
@@ -17,12 +18,12 @@ this crate.
   runtime crates.
 - Prefer `bitfun-core-types` for shared DTOs and `bitfun-runtime-ports` for
   cross-layer traits.
-- Keep the default feature lightweight; feature groups such as search, LSP,
-  cron, or snapshot should not become new crates until measured compile cost
-  proves the split is needed.
-- LSP manifest and protocol DTOs belong in `bitfun-core-types`; plugin package
-  filesystem IO and process lifecycle stay in the core compatibility adapter or
-  a reviewed concrete service owner.
+- Keep dependency features explicit. Non-LSP consumers should use
+  `default-features = false`; LSP consumers must enable the `lsp` feature.
+- LSP manifest and protocol DTOs belong in `bitfun-core-types`; reusable LSP
+  package, protocol, detection, debounce, watch, and process-manager helpers
+  belong in `services-core`; product workspace state, event emission, global
+  singletons, and file-sync orchestration stay outside this crate.
 - Runtime call sites that touch agent execution, scheduler state, workspace
   managers, filesystem orchestration, or product behavior stay in core until a
   reviewed port/provider design and equivalence tests exist.
@@ -34,7 +35,7 @@ this crate.
 ## Verification
 
 ```bash
-cargo test -p bitfun-services-core
+cargo test -p bitfun-services-core --features lsp
 node scripts/check-core-boundaries.mjs
 cargo check -p bitfun-core --features product-full
 ```
