+++ src/crates/services/services-core/Cargo.toml
@@ -10,14 +10,17 @@ name = "bitfun_services_core"
 crate-type = ["rlib"]
 
 [dependencies]
+anyhow = { workspace = true, optional = true }
 bitfun-core-types = { path = "../../contracts/core-types" }
 tokio = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 base64 = { workspace = true }
 chrono = { workspace = true }
+zip = { workspace = true, optional = true }
 thiserror = { workspace = true }
 log = { workspace = true }
+notify = { workspace = true, optional = true }
 ignore = { workspace = true }
 sha2 = { workspace = true }
 which = { workspace = true }
@@ -26,6 +29,10 @@ regex = { workspace = true }
 
 [target.'cfg(windows)'.dependencies]
 win32job = { workspace = true }
+
+[features]
+default = ["lsp"]
+lsp = ["dep:anyhow", "dep:notify", "dep:zip"]
 
 [dev-dependencies]
 tempfile = { workspace = true }
