+++ src/crates/services/services-integrations/Cargo.toml
@@ -23,7 +23,7 @@ aes-gcm = { workspace = true, optional = true }
 anyhow = { workspace = true, optional = true }
 async-trait = { workspace = true, optional = true }
 base64 = { workspace = true, optional = true }
-bitfun-services-core = { path = "../services-core", optional = true }
+bitfun-services-core = { path = "../services-core", default-features = false, optional = true }
 chrono = { workspace = true, optional = true }
 dunce = { workspace = true, optional = true }
 futures = { workspace = true, optional = true }
